Successfully displaying the main fragment
