+++ Ticketing/.idea/misc.xml
@@ -15,6 +15,7 @@
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/activity_search.xml" value="0.1" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/button_shadow.xml" value="0.23550724637681159" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/closed_ticket.xml" value="0.1" />
+        <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/fragment_main.xml" value="0.1171875" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/ticket.xml" value="0.1" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/menu/main.xml" value="0.1171875" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1245" />
