+++ Ticketing/app/src/main/res/values/strings.xml
@@ -13,4 +13,6 @@
     <string name="search_button">Search</string>
 
     <string name="fab_content_desc">Send new support ticket</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
